pyproject.toml
[project]
name = "metalearn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "higher>=0.2.1",
    "jupyter>=1.1.1",
    "peft>=0.16.0",
    "python-dotenv>=1.1.1",
    "torch>=2.7.1",
    "transformers>=4.53.1",
]
